pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-pagination"
version = "0.1.0"
description = "A generic, reusable pagination view for discord.py bots."
requires-python = ">=3.10"
dependencies = ["discord.py>=2.0"]

[project.optional-dependencies]
dev = ["pyright"]
